+++ realm/realm-library/src/main/java/io/realm/RealmQuery.java
@@ -291,8 +291,7 @@ public final class RealmQuery<E extends RealmModel> {
      * @param fieldName the field to compare.
      * @param value the value to compare with.
      * @return the query object.
-     * @throws java.lang.IllegalArgumentException if one or more arguments do not match class or field type, or if
-     * {@code fieldName} points to a linked field.
+     * @throws java.lang.IllegalArgumentException if one or more arguments do not match class or field type.
      */
     public RealmQuery<E> equalTo(String fieldName, byte[] value) {
         long[] columnIndices = schema.getColumnIndices(fieldName, RealmFieldType.BINARY);
@@ -667,8 +666,7 @@ public final class RealmQuery<E extends RealmModel> {
      * @param fieldName the field to compare.
      * @param value the value to compare with.
      * @return the query object.
-     * @throws java.lang.IllegalArgumentException if one or more arguments do not match class or field type, or if
-     * {@code fieldName} points to a linked field.
+     * @throws java.lang.IllegalArgumentException if one or more arguments do not match class or field type.
      */
     public RealmQuery<E> notEqualTo(String fieldName, byte[] value) {
         long[] columnIndices = schema.getColumnIndices(fieldName, RealmFieldType.BINARY);
@@ -1198,21 +1196,6 @@ public final class RealmQuery<E extends RealmModel> {
     }
 
     /**
-     * Condition that value of field contains the specified byte array.
-     *
-     * @param fieldName the field to compare.
-     * @param value the byte array.
-     * @return the query object.
-     * @throws java.lang.IllegalArgumentException if one or more arguments do not match class or field type, or if
-     * {@code fieldName} points to a linked field.
-     */
-    public RealmQuery<E> contains(String fieldName, byte[] value) {
-        long columnIndices[] = schema.getColumnIndices(fieldName, RealmFieldType.BINARY);
-        this.query.contains(columnIndices, value);
-        return this;
-    }
-
-    /**
      * Condition that the value of field begins with the specified string.
      *
      * @param fieldName the field to compare.
@@ -1240,21 +1223,6 @@ public final class RealmQuery<E extends RealmModel> {
     }
 
     /**
-     * Condition that the value of field begins with the specified byte array.
-     *
-     * @param fieldName the field to compare.
-     * @param value the byte array.
-     * @return the query object.
-     * @throws java.lang.IllegalArgumentException if one or more arguments do not match class or field type, or if
-     * {@code fieldName} points to a linked field.
-     */
-    public RealmQuery<E> beginsWith(String fieldName, byte[] value) {
-        long columnIndices[] = schema.getColumnIndices(fieldName, RealmFieldType.BINARY);
-        this.query.beginsWith(columnIndices, value);
-        return this;
-    }
-
-    /**
      * Condition that the value of field ends with the specified string.
      *
      * @param fieldName the field to compare.
@@ -1278,21 +1246,6 @@ public final class RealmQuery<E extends RealmModel> {
     public RealmQuery<E> endsWith(String fieldName, String value, Case casing) {
         long columnIndices[] = schema.getColumnIndices(fieldName, RealmFieldType.STRING);
         this.query.endsWith(columnIndices, value, casing);
-        return this;
-    }
-
-    /**
-     * Condition that the value of field ends with the specified byte array.
-     *
-     * @param fieldName the field to compare.
-     * @param value the byte array.
-     * @return the query object.
-     * @throws java.lang.IllegalArgumentException if one or more arguments do not match class or field type, or if
-     * {@code fieldName} points to a linked field.
-     */
-    public RealmQuery<E> endsWith(String fieldName, byte[] value) {
-        long columnIndices[] = schema.getColumnIndices(fieldName, RealmFieldType.BINARY);
-        this.query.endsWith(columnIndices, value);
         return this;
     }